rename uecs_device.go
